refactor: restore top-level estimation layout
