+++ src/estimation/mod.rs
@@ -1,0 +1,6 @@
+//! Pose estimation algorithms and estimator modules.
+
+pub mod map_projection;
+pub mod two_view;
+
+pub use map_projection::MapProjectionEstimator;
