+++ src/estimation/two_view.rs
@@ -1,4 +1,4 @@
-//! Bootstrap-stage logic for monocular tracking using two-view triangulation.a
+//! Two-view geometric initialization for monocular tracking.
 //              (3D point)
 //                  X
 //                /   \
@@ -16,29 +16,33 @@ use kornia_3d::camera::PinholeCamera;
 use kornia_3d::pose::Pose3d;
 use kornia_3d::pose::{TwoViewConfig, TwoViewModel, two_view_estimate};
 use kornia_algebra::Vec3F64;
-use kornia_imgproc::features::OrbMatchConfig;
+use kornia_imgproc::features::{OrbMatchConfig, match_orb_descriptors};
 
-use crate::odometry::estimation::matching::match_orb_descriptors;
 use crate::frame::OrbFeatures;
 
-/// Configuration for the bootstrap stage.
+/// Configuration for two-view initialization.
 #[derive(Debug, Clone)]
-pub struct BootstrapConfig {
-    /// ORB descriptor matcher settings.
-    pub match_config: OrbMatchConfig,
-    /// Two-view model estimation settings.
-    pub two_view_config: TwoViewConfig,
+pub struct TwoViewAcceptanceConfig {
     /// Minimum descriptor matches required before two-view estimation.
     pub min_matches: usize,
-    /// Minimum inlier count required to accept bootstrap.
+    /// Minimum inlier count required to accept initialization.
     pub min_inliers: usize,
-    /// Minimum median parallax in degrees required to accept bootstrap.
-    pub min_parallax_deg: f64,
-    /// Minimum triangulated points required to accept bootstrap.
+    /// Minimum triangulated points required to accept initialization.
     pub min_triangulated: usize,
 }
 
-impl Default for BootstrapConfig {
+/// Configuration for two-view initialization.
+#[derive(Debug, Clone)]
+pub struct TwoViewInitConfig {
+    /// ORB descriptor matcher settings.
+    pub match_config: OrbMatchConfig,
+    /// Two-view model estimation settings.
+    pub estimation_config: TwoViewConfig,
+    /// Acceptance thresholds applied on top of the estimator result.
+    pub acceptance_config: TwoViewAcceptanceConfig,
+}
+
+impl Default for TwoViewInitConfig {
     fn default() -> Self {
         Self {
             match_config: OrbMatchConfig {
@@ -47,18 +51,19 @@ impl Default for BootstrapConfig {
                 check_orientation: true,
                 histo_length: 30,
             },
-            two_view_config: TwoViewConfig::default(),
-            min_matches: 100,
-            min_inliers: 30,
-            min_parallax_deg: 1.0,
-            min_triangulated: 50,
+            estimation_config: TwoViewConfig::default(),
+            acceptance_config: TwoViewAcceptanceConfig {
+                min_matches: 100,
+                min_inliers: 30,
+                min_triangulated: 50,
+            },
         }
     }
 }
 
-/// Bootstrap rejection reason.
+/// Two-view initialization rejection reason.
 #[derive(Debug, Clone, Copy, PartialEq, Eq, Hash)]
-pub enum BootstrapRejectReason {
+pub enum TwoViewInitRejectReason {
     /// Not enough descriptor matches to run two-view estimation.
     LowMatches,
     /// Two-view estimation failed.
@@ -73,15 +78,15 @@ pub enum BootstrapRejectReason {
     LowParallax,
 }
 
-/// Result of one bootstrap step.
+/// Result of one two-view initialization attempt.
 #[derive(Debug, Clone)]
-pub enum BootstrapOutcome {
-    /// Attempt was rejected by bootstrap gates.
+pub enum TwoViewInitOutcome {
+    /// Attempt was rejected by two-view initialization gates.
     Rejected {
-        /// Why the bootstrap attempt failed.
-        reason: BootstrapRejectReason,
+        /// Why the two-view initialization attempt failed.
+        reason: TwoViewInitRejectReason,
     },
-    /// Bootstrap succeeded.
+    /// Two-view initialization succeeded.
     Initialized {
         /// New world-to-camera pose for the current frame.
         pose_world_to_cam: Pose3d,
@@ -98,15 +103,17 @@ pub enum BootstrapOutcome {
     },
 }
 
-/// Attempt two-view bootstrap between a reference frame and the current frame.
-pub fn try_bootstrap(
+/// Attempt two-view initialization between a reference frame and the current frame.
+pub fn try_initialize_two_view(
     ref_features: &OrbFeatures,
     ref_pose: &Pose3d,
     curr_features: &OrbFeatures,
     curr_pose: &Pose3d,
     camera: &PinholeCamera,
-    config: &BootstrapConfig,
-) -> BootstrapOutcome {
+    config: &TwoViewInitConfig,
+) -> TwoViewInitOutcome {
+    let acceptance = &config.acceptance_config;
+
     let matches = match_orb_descriptors(
         &ref_features.orientations,
         &ref_features.descriptors,
@@ -114,9 +121,9 @@ pub fn try_bootstrap(
         &curr_features.descriptors,
         config.match_config,
     );
-    if matches.len() < config.min_matches {
-        return BootstrapOutcome::Rejected {
-            reason: BootstrapRejectReason::LowMatches,
+    if matches.len() < acceptance.min_matches {
+        return TwoViewInitOutcome::Rejected {
+            reason: TwoViewInitRejectReason::LowMatches,
         };
     }
 
@@ -132,35 +139,35 @@ pub fn try_bootstrap(
         &current_pts,
         &k,
         &k,
-        &config.two_view_config,
+        &config.estimation_config,
     ) else {
-        return BootstrapOutcome::Rejected {
-            reason: BootstrapRejectReason::EstimationFailed,
+        return TwoViewInitOutcome::Rejected {
+            reason: TwoViewInitRejectReason::EstimationFailed,
         };
     };
 
     if !matches!(result.model, TwoViewModel::Fundamental(_)) {
-        return BootstrapOutcome::Rejected {
-            reason: BootstrapRejectReason::WrongModel,
+        return TwoViewInitOutcome::Rejected {
+            reason: TwoViewInitRejectReason::WrongModel,
         };
     }
 
-    if result.points3d.len() < config.min_triangulated {
-        return BootstrapOutcome::Rejected {
-            reason: BootstrapRejectReason::LowTriangulated,
+    if result.points3d.len() < acceptance.min_triangulated {
+        return TwoViewInitOutcome::Rejected {
+            reason: TwoViewInitRejectReason::LowTriangulated,
         };
     }
 
-    if result.inlier_indices.len() < config.min_inliers {
-        return BootstrapOutcome::Rejected {
-            reason: BootstrapRejectReason::LowInliers,
+    if result.inlier_indices.len() < acceptance.min_inliers {
+        return TwoViewInitOutcome::Rejected {
+            reason: TwoViewInitRejectReason::LowInliers,
         };
     }
 
     let median_parallax_deg = result.median_parallax_deg(&reference_pts, &current_pts, camera);
-    if median_parallax_deg < config.min_parallax_deg {
-        return BootstrapOutcome::Rejected {
-            reason: BootstrapRejectReason::LowParallax,
+    if median_parallax_deg < config.estimation_config.min_parallax_deg {
+        return TwoViewInitOutcome::Rejected {
+            reason: TwoViewInitRejectReason::LowParallax,
         };
     }
 
@@ -182,7 +189,7 @@ pub fn try_bootstrap(
     );
     let motion_increment = Pose3d::between(curr_pose, &new_pose);
 
-    BootstrapOutcome::Initialized {
+    TwoViewInitOutcome::Initialized {
         pose_world_to_cam: new_pose,
         motion_increment,
         matches,
