+++ src/lib.rs
@@ -1,9 +1,10 @@
 //! Visual odometry and SLAM building blocks for kornia-rs.
 
 pub mod frame;
+pub mod estimation;
 pub mod mapping;
 pub mod odometry;
 
 pub use frame::{Frame, OrbFeatures};
-pub use odometry::{Estimator, OdometryMode, OdometryResult, OdometryState, OdometryStatus};
-pub use odometry::estimation::MapProjectionEstimator;
+pub use estimation::MapProjectionEstimator;
+pub use odometry::{OdometryMode, OdometryResult, OdometryState, OdometryStatus};
